Restructured AudioAnalyser folder
